Added Hindi Language Translation File
I would like to contribute by adding a Hindi translation file to this project. Hindi is the national language of India and is spoken by over 500 million people worldwide. By adding a Hindi translation file, I can help to make this project more accessible to a wider audience.
I believe that adding a Hindi translation file to this project would be a valuable contribution. It would make the project more accessible to a wider audience.